p-nomr of a vector
The p-norm of a vector v = (v 1 ,v 2 ,...,v n ) in n-dimensional space is defined as:    ||V|| = root(P)√v1^p + v2^p.... + vn^p.

for the special case of p = 2, this results in the traditional Euclidean norm,which represents the length of the vector. For example, the Euclidean norm of a two-dimensional vector with coordinates (4,3) has a Euclidean norm of √ 4 2 +3 2 = √ 16+9 = √ 25 = 5. Give an implementation of a method named norm such that norm(v, p) returns the p-normvalue of v and norm(v) returns the Euclidean norm of v, where v is represented as an array of coordinates.